@@ -16,6 +16,8 @@ class _TabInfo {
 }
 
 class CupertinoTabBarDemo extends StatelessWidget {
+  const CupertinoTabBarDemo();
+
   @override
   Widget build(BuildContext context) {
     final _tabInfo = [
@@ -72,7 +74,7 @@ class _CupertinoDemoTab extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return CupertinoPageScaffold(
-      navigationBar: CupertinoNavigationBar(),
+      navigationBar: const CupertinoNavigationBar(),
       backgroundColor: CupertinoColors.systemBackground,
       child: Center(
         child: Icon(
