@@ -70,7 +70,7 @@ class _MenuDemoState extends State<MenuDemo> {
         automaticallyImplyLeading: false,
       ),
       body: Padding(
-        padding: EdgeInsets.symmetric(horizontal: 20),
+        padding: const EdgeInsets.symmetric(horizontal: 20),
         child: Center(
           child: demo,
         ),
