@@ -12,6 +12,8 @@ import 'package:gallery/l10n/gallery_localizations.dart';
 // BEGIN textFieldDemo
 
 class TextFieldDemo extends StatelessWidget {
+  const TextFieldDemo();
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -19,7 +21,7 @@ class TextFieldDemo extends StatelessWidget {
         automaticallyImplyLeading: false,
         title: Text(GalleryLocalizations.of(context).demoTextFieldTitle),
       ),
-      body: TextFormFieldDemo(),
+      body: const TextFormFieldDemo(),
     );
   }
 }
@@ -245,7 +247,7 @@ class TextFormFieldDemoState extends State<TextFormFieldDemo> {
                 TextFormField(
                   cursorColor: cursorColor,
                   decoration: InputDecoration(
-                    border: OutlineInputBorder(),
+                    border: const OutlineInputBorder(),
                     hintText: GalleryLocalizations.of(context)
                         .demoTextFieldTellUsAboutYourself,
                     helperText: GalleryLocalizations.of(context)
@@ -260,7 +262,7 @@ class TextFormFieldDemoState extends State<TextFormFieldDemo> {
                   cursorColor: cursorColor,
                   keyboardType: TextInputType.number,
                   decoration: InputDecoration(
-                    border: OutlineInputBorder(),
+                    border: const OutlineInputBorder(),
                     labelText:
                         GalleryLocalizations.of(context).demoTextFieldSalary,
                     suffixText:
