@@ -34,105 +34,85 @@ const _horizontalDesktopPadding = 81.0;
 const _carouselHeightMin = 200.0 + 2 * _carouselItemMargin;
 const _desktopCardsPerPage = 4;
 
-const _shrineTitle = 'Shrine';
-const _rallyTitle = 'Rally';
-const _craneTitle = 'Crane';
-const _homeCategoryMaterial = 'MATERIAL';
-const _homeCategoryCupertino = 'CUPERTINO';
-
 class ToggleSplashNotification extends Notification {}
-
-class NavigatorKeys {
-  static final shrine = GlobalKey<NavigatorState>();
-  static final rally = GlobalKey<NavigatorState>();
-  static final crane = GlobalKey<NavigatorState>();
-  static final fortnightly = GlobalKey<NavigatorState>();
-  static final starter = GlobalKey<NavigatorState>();
-}
 
 class HomePage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     var carouselHeight = _carouselHeight(.7, context);
     final isDesktop = isDisplayDesktop(context);
+    final localizations = GalleryLocalizations.of(context);
+    final studyDemos = studies(localizations);
     final carouselCards = <Widget>[
       _CarouselCard(
-        title: _shrineTitle,
-        subtitle: GalleryLocalizations.of(context).shrineDescription,
-        asset: AssetImage('assets/studies/shrine_card.png'),
+        demo: studyDemos['shrine'],
+        asset: const AssetImage('assets/studies/shrine_card.png'),
         assetColor: const Color(0xFFFEDBD0),
-        assetDark: AssetImage('assets/studies/shrine_card_dark.png'),
+        assetDark: const AssetImage('assets/studies/shrine_card_dark.png'),
         assetDarkColor: const Color(0xFF543B3C),
         textColor: shrineBrown900,
-        study: ShrineApp(navigatorKey: NavigatorKeys.shrine),
-        navigatorKey: NavigatorKeys.shrine,
+        studyRoute: ShrineApp.loginRoute,
       ),
       _CarouselCard(
-        title: _rallyTitle,
-        subtitle: GalleryLocalizations.of(context).rallyDescription,
+        demo: studyDemos['rally'],
         textColor: RallyColors.accountColors[0],
-        asset: AssetImage('assets/studies/rally_card.png'),
+        asset: const AssetImage('assets/studies/rally_card.png'),
         assetColor: const Color(0xFFD1F2E6),
-        assetDark: AssetImage('assets/studies/rally_card_dark.png'),
+        assetDark: const AssetImage('assets/studies/rally_card_dark.png'),
         assetDarkColor: const Color(0xFF253538),
-        study: RallyApp(navigatorKey: NavigatorKeys.rally),
-        navigatorKey: NavigatorKeys.rally,
+        studyRoute: RallyApp.loginRoute,
       ),
       _CarouselCard(
-        title: _craneTitle,
-        subtitle: GalleryLocalizations.of(context).craneDescription,
-        asset: AssetImage('assets/studies/crane_card.png'),
+        demo: studyDemos['crane'],
+        asset: const AssetImage('assets/studies/crane_card.png'),
         assetColor: const Color(0xFFFBF6F8),
-        assetDark: AssetImage('assets/studies/crane_card_dark.png'),
+        assetDark: const AssetImage('assets/studies/crane_card_dark.png'),
         assetDarkColor: const Color(0xFF591946),
         textColor: cranePurple700,
-        study: CraneApp(navigatorKey: NavigatorKeys.crane),
-        navigatorKey: NavigatorKeys.crane,
+        studyRoute: CraneApp.defaultRoute,
       ),
       _CarouselCard(
-        title: fortnightlyTitle,
-        subtitle: GalleryLocalizations.of(context).fortnightlyDescription,
-        asset: AssetImage('assets/studies/fortnightly_card.png'),
+        demo: studyDemos['fortnightly'],
+        asset: const AssetImage('assets/studies/fortnightly_card.png'),
         assetColor: Colors.white,
-        assetDark: AssetImage('assets/studies/fortnightly_card_dark.png'),
+        assetDark: const AssetImage('assets/studies/fortnightly_card_dark.png'),
         assetDarkColor: const Color(0xFF1F1F1F),
-        study: FortnightlyApp(navigatorKey: NavigatorKeys.fortnightly),
-        navigatorKey: NavigatorKeys.fortnightly,
+        studyRoute: FortnightlyApp.defaultRoute,
       ),
       _CarouselCard(
-        title: GalleryLocalizations.of(context).starterAppTitle,
-        subtitle: GalleryLocalizations.of(context).starterAppDescription,
-        asset: AssetImage('assets/studies/starter_card.png'),
+        demo: studyDemos['starterApp'],
+        asset: const AssetImage('assets/studies/starter_card.png'),
         assetColor: const Color(0xFFFAF6FE),
-        assetDark: AssetImage('assets/studies/starter_card_dark.png'),
+        assetDark: const AssetImage('assets/studies/starter_card_dark.png'),
         assetDarkColor: const Color(0xFF3F3D45),
         textColor: Colors.black,
-        study: StarterApp(navigatorKey: NavigatorKeys.starter),
-        navigatorKey: NavigatorKeys.starter,
+        studyRoute: StarterApp.defaultRoute,
       ),
     ];
 
     if (isDesktop) {
       final desktopCategoryItems = <_DesktopCategoryItem>[
         _DesktopCategoryItem(
-          title: _homeCategoryMaterial,
-          asset: AssetImage('assets/icons/material/material.png'),
-          demos: materialDemos(context),
+          category: GalleryDemoCategory.material,
+          asset: const AssetImage('assets/icons/material/material.png'),
+          demos: materialDemos(localizations),
         ),
         _DesktopCategoryItem(
-          title: _homeCategoryCupertino,
-          asset: AssetImage('assets/icons/cupertino/cupertino.png'),
-          demos: cupertinoDemos(context),
+          category: GalleryDemoCategory.cupertino,
+          asset: const AssetImage('assets/icons/cupertino/cupertino.png'),
+          demos: cupertinoDemos(localizations),
         ),
         _DesktopCategoryItem(
-          title: GalleryLocalizations.of(context).homeCategoryReference,
-          asset: AssetImage('assets/icons/reference/reference.png'),
-          demos: referenceDemos(context),
+          category: GalleryDemoCategory.other,
+          asset: const AssetImage('assets/icons/reference/reference.png'),
+          demos: otherDemos(localizations),
         ),
       ];
 
       return Scaffold(
         body: ListView(
+          // Makes integration tests possible.
+          key: const ValueKey('HomeListView'),
           padding: EdgeInsetsDirectional.only(
             top: isDesktop ? firstHeaderDesktopTopPadding : 21,
           ),
@@ -155,8 +135,8 @@ class HomePage extends StatelessWidget {
             /// the container which is why we created this Focus variable.
             Focus(
               focusNode:
-                  InheritedBackdropFocusNodes.of(context).backLayerFocusNode,
-              child: SizedBox(),
+                  InheritedBackdropFocusNodes.of(context).homePageFocusNode,
+              child: const SizedBox(),
             ),
             Container(
               height: carouselHeight,
@@ -191,9 +171,10 @@ class HomePage extends StatelessWidget {
                   FadeInImagePlaceholder(
                     image: Theme.of(context).colorScheme.brightness ==
                             Brightness.dark
-                        ? AssetImage('assets/logo/flutter_logo.png')
-                        : AssetImage('assets/logo/flutter_logo_color.png'),
-                    placeholder: SizedBox.shrink(),
+                        ? const AssetImage('assets/logo/flutter_logo.png')
+                        : const AssetImage(
+                            'assets/logo/flutter_logo_color.png'),
+                    placeholder: const SizedBox.shrink(),
                     excludeFromSemantics: true,
                   ),
                   Expanded(
@@ -305,7 +286,7 @@ class _AnimatedHomePageState extends State<_AnimatedHomePage>
     super.initState();
     _animationController = AnimationController(
       vsync: this,
-      duration: Duration(milliseconds: 800),
+      duration: const Duration(milliseconds: 800),
     );
 
     if (widget.isSplashPageAnimationFinished) {
@@ -336,13 +317,17 @@ class _AnimatedHomePageState extends State<_AnimatedHomePage>
 
   @override
   Widget build(BuildContext context) {
+    final localizations = GalleryLocalizations.of(context);
     return Stack(
       children: [
         ListView(
+          // Makes integration tests possible.
+          key: const ValueKey('HomeListView'),
           children: [
-            SizedBox(height: 8),
+            const SizedBox(height: 8),
             Container(
-              margin: EdgeInsets.symmetric(horizontal: _horizontalPadding),
+              margin:
+                  const EdgeInsets.symmetric(horizontal: _horizontalPadding),
               child: ExcludeSemantics(child: _GalleryHeader()),
             ),
             _Carousel(
@@ -350,34 +335,44 @@ class _AnimatedHomePageState extends State<_AnimatedHomePage>
               animationController: _animationController,
             ),
             Container(
-              margin: EdgeInsets.symmetric(horizontal: _horizontalPadding),
+              margin:
+                  const EdgeInsets.symmetric(horizontal: _horizontalPadding),
               child: _CategoriesHeader(),
             ),
             _AnimatedCategoryItem(
               startDelayFraction: 0.00,
               controller: _animationController,
               child: CategoryListItem(
-                title: _homeCategoryMaterial,
+                key: const PageStorageKey<GalleryDemoCategory>(
+                  GalleryDemoCategory.material,
+                ),
+                category: GalleryDemoCategory.material,
                 imageString: 'assets/icons/material/material.png',
-                demos: materialDemos(context),
+                demos: materialDemos(localizations),
               ),
             ),
             _AnimatedCategoryItem(
               startDelayFraction: 0.05,
               controller: _animationController,
               child: CategoryListItem(
-                title: _homeCategoryCupertino,
+                key: const PageStorageKey<GalleryDemoCategory>(
+                  GalleryDemoCategory.cupertino,
+                ),
+                category: GalleryDemoCategory.cupertino,
                 imageString: 'assets/icons/cupertino/cupertino.png',
-                demos: cupertinoDemos(context),
+                demos: cupertinoDemos(localizations),
               ),
             ),
             _AnimatedCategoryItem(
               startDelayFraction: 0.10,
               controller: _animationController,
               child: CategoryListItem(
-                title: GalleryLocalizations.of(context).homeCategoryReference,
+                key: const PageStorageKey<GalleryDemoCategory>(
+                  GalleryDemoCategory.other,
+                ),
+                category: GalleryDemoCategory.other,
                 imageString: 'assets/icons/reference/reference.png',
-                demos: referenceDemos(context),
+                demos: otherDemos(localizations),
               ),
             ),
           ],
@@ -406,12 +401,12 @@ class _AnimatedHomePageState extends State<_AnimatedHomePage>
 
 class _DesktopCategoryItem extends StatelessWidget {
   const _DesktopCategoryItem({
-    this.title,
+    this.category,
     this.asset,
     this.demos,
   });
 
-  final String title;
+  final GalleryDemoCategory category;
   final ImageProvider asset;
   final List<GalleryDemo> demos;
 
@@ -429,7 +424,7 @@ class _DesktopCategoryItem extends StatelessWidget {
           child: Column(
             children: [
               _DesktopCategoryHeader(
-                title: title,
+                category: category,
                 asset: asset,
               ),
               Divider(
@@ -443,13 +438,15 @@ class _DesktopCategoryItem extends StatelessWidget {
                   removeTop: true,
                   context: context,
                   child: ListView(
+                    // Makes integration tests possible.
+                    key: ValueKey('${category.name}DemoList'),
                     children: [
                       const SizedBox(height: 12),
                       for (GalleryDemo demo in demos)
                         CategoryDemoItem(
                           demo: demo,
                         ),
-                      SizedBox(height: 12),
+                      const SizedBox(height: 12),
                     ],
                   ),
                 ),
@@ -464,24 +461,26 @@ class _DesktopCategoryItem extends StatelessWidget {
 
 class _DesktopCategoryHeader extends StatelessWidget {
   const _DesktopCategoryHeader({
-    this.title,
+    this.category,
     this.asset,
   });
-  final String title;
+  final GalleryDemoCategory category;
   final ImageProvider asset;
 
   @override
   Widget build(BuildContext context) {
     ColorScheme colorScheme = Theme.of(context).colorScheme;
     return Material(
+      // Makes integration tests possible.
+      key: ValueKey('${category.name}CategoryHeader'),
       color: colorScheme.onBackground,
       child: Row(
         children: [
           Padding(
-            padding: EdgeInsets.all(10),
+            padding: const EdgeInsets.all(10),
             child: FadeInImagePlaceholder(
               image: asset,
-              placeholder: SizedBox(
+              placeholder: const SizedBox(
                 height: 64,
                 width: 64,
               ),
@@ -492,11 +491,11 @@ class _DesktopCategoryHeader extends StatelessWidget {
           ),
           Flexible(
             child: Padding(
-              padding: EdgeInsetsDirectional.only(start: 8),
+              padding: const EdgeInsetsDirectional.only(start: 8),
               child: Semantics(
                 header: true,
                 child: Text(
-                  title,
+                  category.displayTitle(GalleryLocalizations.of(context)),
                   style: Theme.of(context).textTheme.headline5.apply(
                         color: colorScheme.onSurface,
                       ),
@@ -810,7 +809,7 @@ class _DesktopCarouselState extends State<_DesktopCarousel> {
           ),
           child: ListView.builder(
             scrollDirection: Axis.horizontal,
-            physics: _SnappingScrollPhysics(),
+            physics: const _SnappingScrollPhysics(),
             controller: _controller,
             itemExtent: itemWidth,
             itemCount: widget.children.length,
@@ -822,7 +821,7 @@ class _DesktopCarouselState extends State<_DesktopCarousel> {
             onTap: () {
               _controller.animateTo(
                 _controller.offset - itemWidth,
-                duration: Duration(milliseconds: 200),
+                duration: const Duration(milliseconds: 200),
                 curve: Curves.easeInOut,
               );
             },
@@ -833,7 +832,7 @@ class _DesktopCarouselState extends State<_DesktopCarousel> {
             onTap: () {
               _controller.animateTo(
                 _controller.offset + itemWidth,
-                duration: Duration(milliseconds: 200),
+                duration: const Duration(milliseconds: 200),
                 curve: Curves.easeInOut,
               );
             },
@@ -928,7 +927,7 @@ class _DesktopPageButton extends StatelessWidget {
               : MaterialLocalizations.of(context).previousPageTooltip,
           child: Material(
             color: Colors.black.withOpacity(0.5),
-            shape: CircleBorder(),
+            shape: const CircleBorder(),
             clipBehavior: Clip.antiAlias,
             child: InkWell(
               onTap: onTap,
@@ -947,26 +946,22 @@ class _DesktopPageButton extends StatelessWidget {
 class _CarouselCard extends StatelessWidget {
   const _CarouselCard({
     Key key,
-    this.title,
-    this.subtitle,
+    this.demo,
     this.asset,
     this.assetDark,
     this.assetColor,
     this.assetDarkColor,
     this.textColor,
-    this.study,
-    this.navigatorKey,
+    this.studyRoute,
   }) : super(key: key);
 
-  final String title;
-  final String subtitle;
+  final GalleryDemo demo;
   final ImageProvider asset;
   final ImageProvider assetDark;
   final Color assetColor;
   final Color assetDarkColor;
   final Color textColor;
-  final Widget study;
-  final GlobalKey<NavigatorState> navigatorKey;
+  final String studyRoute;
 
   @override
   Widget build(BuildContext context) {
@@ -977,23 +972,17 @@ class _CarouselCard extends StatelessWidget {
     final textColor = isDark ? Colors.white.withOpacity(0.87) : this.textColor;
 
     return Container(
+      // Makes integration tests possible.
+      key: ValueKey(demo.describe),
       margin:
           EdgeInsets.all(isDisplayDesktop(context) ? 0 : _carouselItemMargin),
       child: Material(
         elevation: 4,
         shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(10)),
         clipBehavior: Clip.antiAlias,
-        color: Colors.grey,
         child: InkWell(
           onTap: () {
-            Navigator.of(context).push<void>(
-              MaterialPageRoute(
-                builder: (context) => _StudyWrapper(
-                  study: study,
-                  navigatorKey: navigatorKey,
-                ),
-              ),
-            );
+            Navigator.of(context).pushNamed(studyRoute);
           },
           child: Stack(
             fit: StackFit.expand,
@@ -1010,19 +999,19 @@ class _CarouselCard extends StatelessWidget {
                   ),
                 ),
               Padding(
-                padding: EdgeInsetsDirectional.fromSTEB(16, 0, 16, 16),
+                padding: const EdgeInsetsDirectional.fromSTEB(16, 0, 16, 16),
                 child: Column(
                   crossAxisAlignment: CrossAxisAlignment.start,
                   mainAxisAlignment: MainAxisAlignment.end,
                   children: [
                     Text(
-                      title,
+                      demo.title,
                       style: textTheme.caption.apply(color: textColor),
                       maxLines: 3,
                       overflow: TextOverflow.visible,
                     ),
                     Text(
-                      subtitle,
+                      demo.subtitle,
                       style: textTheme.overline.apply(color: textColor),
                       maxLines: 5,
                       overflow: TextOverflow.visible,
@@ -1046,136 +1035,59 @@ double _carouselHeight(double scaleFactor, BuildContext context) => math.max(
 
 /// Wrap the studies with this to display a back button and allow the user to
 /// exit them at any time.
-class _StudyWrapper extends StatefulWidget {
-  const _StudyWrapper({
+class StudyWrapper extends StatefulWidget {
+  const StudyWrapper({
     Key key,
     this.study,
-    this.navigatorKey,
   }) : super(key: key);
 
   final Widget study;
-  final GlobalKey<NavigatorState> navigatorKey;
 
   @override
   _StudyWrapperState createState() => _StudyWrapperState();
 }
 
-class _StudyWrapperState extends State<_StudyWrapper> {
-  FocusNode backButtonFocusNode;
-
-  @override
-  void initState() {
-    super.initState();
-    backButtonFocusNode = FocusNode();
-  }
-
-  @override
-  void dispose() {
-    backButtonFocusNode.dispose();
-    super.dispose();
-  }
-
+class _StudyWrapperState extends State<StudyWrapper> {
   @override
   Widget build(BuildContext context) {
     final colorScheme = Theme.of(context).colorScheme;
     final textTheme = Theme.of(context).textTheme;
     return ApplyTextOptions(
-      child: FocusTraversalGroup(
-        policy: StudyWrapperFocusTraversalPolicy(
-          backButtonFocusNode: backButtonFocusNode,
-          studyNavigatorKey: widget.navigatorKey,
-        ),
-        child: InheritedFocusNodes(
-          backButtonFocusNode: backButtonFocusNode,
-          child: Stack(
-            children: [
-              Semantics(
-                sortKey: const OrdinalSortKey(1),
-                child: widget.study,
-              ),
-              Align(
-                alignment: AlignmentDirectional.bottomStart,
-                child: Padding(
-                  padding: const EdgeInsets.all(16),
-                  child: Semantics(
-                    sortKey: const OrdinalSortKey(0),
-                    label: GalleryLocalizations.of(context).backToGallery,
-                    button: true,
-                    excludeSemantics: true,
-                    child: FloatingActionButton.extended(
-                      focusNode: backButtonFocusNode,
-                      onPressed: () {
-                        Navigator.of(context).pop();
-                      },
-                      icon: IconTheme(
-                        data: IconThemeData(color: colorScheme.onPrimary),
-                        child: BackButtonIcon(),
-                      ),
-                      label: Text(
-                        MaterialLocalizations.of(context).backButtonTooltip,
-                        style: textTheme.button
-                            .apply(color: colorScheme.onPrimary),
-                      ),
-                    ),
+      child: Stack(
+        children: [
+          Semantics(
+            sortKey: const OrdinalSortKey(1),
+            child: widget.study,
+          ),
+          Align(
+            alignment: AlignmentDirectional.bottomStart,
+            child: Padding(
+              padding: const EdgeInsets.all(16),
+              child: Semantics(
+                sortKey: const OrdinalSortKey(0),
+                label: GalleryLocalizations.of(context).backToGallery,
+                button: true,
+                excludeSemantics: true,
+                child: FloatingActionButton.extended(
+                  key: const ValueKey('Back'),
+                  onPressed: () {
+                    Navigator.of(context)
+                        .popUntil((route) => route.settings.name == '/');
+                  },
+                  icon: IconTheme(
+                    data: IconThemeData(color: colorScheme.onPrimary),
+                    child: const BackButtonIcon(),
+                  ),
+                  label: Text(
+                    MaterialLocalizations.of(context).backButtonTooltip,
+                    style: textTheme.button.apply(color: colorScheme.onPrimary),
                   ),
                 ),
               ),
-            ],
+            ),
           ),
-        ),
+        ],
       ),
     );
-  }
-}
-
-class InheritedFocusNodes extends InheritedWidget {
-  const InheritedFocusNodes({
-    Key key,
-    @required Widget child,
-    @required this.backButtonFocusNode,
-  })  : assert(child != null),
-        super(key: key, child: child);
-
-  final FocusNode backButtonFocusNode;
-
-  static InheritedFocusNodes of(BuildContext context) =>
-      context.dependOnInheritedWidgetOfExactType();
-
-  @override
-  bool updateShouldNotify(InheritedFocusNodes old) => true;
-}
-
-class StudyWrapperFocusTraversalPolicy extends WidgetOrderTraversalPolicy {
-  StudyWrapperFocusTraversalPolicy({
-    @required this.backButtonFocusNode,
-    @required this.studyNavigatorKey,
-  });
-
-  final FocusNode backButtonFocusNode;
-  final GlobalKey<NavigatorState> studyNavigatorKey;
-
-  FocusNode _firstFocusNode() {
-    return studyNavigatorKey.currentState.focusScopeNode.traversalDescendants
-        .toList()
-        .first;
-  }
-
-  @override
-  bool previous(FocusNode currentNode) {
-    if (currentNode == backButtonFocusNode) {
-      return super.previous(_firstFocusNode());
-    } else {
-      return super.previous(currentNode);
-    }
-  }
-
-  @override
-  bool next(FocusNode currentNode) {
-    if (currentNode == backButtonFocusNode) {
-      _firstFocusNode().requestFocus();
-      return true;
-    } else {
-      return super.next(currentNode);
-    }
   }
 }
