@@ -11,57 +11,49 @@ import 'package:gallery/studies/crane/fly_form.dart';
 import 'package:gallery/studies/crane/sleep_form.dart';
 import 'package:gallery/studies/crane/theme.dart';
 
-class CraneApp extends StatefulWidget {
-  const CraneApp({Key key, this.navigatorKey}) : super(key: key);
+class CraneApp extends StatelessWidget {
+  const CraneApp();
 
-  final GlobalKey<NavigatorState> navigatorKey;
+  static const String defaultRoute = '/crane';
 
-  @override
-  _CraneAppState createState() => _CraneAppState();
-}
-
-class _CraneAppState extends State<CraneApp> {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      navigatorKey: widget.navigatorKey,
       title: 'Crane',
       debugShowCheckedModeBanner: false,
       localizationsDelegates: GalleryLocalizations.localizationsDelegates,
       supportedLocales: GalleryLocalizations.supportedLocales,
       locale: GalleryOptions.of(context).locale,
-      initialRoute: '/',
-      onGenerateRoute: _getRoute,
+      initialRoute: CraneApp.defaultRoute,
+      routes: {
+        CraneApp.defaultRoute: (context) => const _Home(),
+      },
       theme: craneTheme.copyWith(
         platform: GalleryOptions.of(context).platform,
       ),
       darkTheme: craneTheme.copyWith(
         platform: GalleryOptions.of(context).platform,
       ),
-      home: ApplyTextOptions(
-        child: Backdrop(
-          frontLayer: Container(),
-          backLayerItems: [
-            FlyForm(index: 0),
-            SleepForm(index: 1),
-            EatForm(index: 2),
-          ],
-          frontTitle: Text('CRANE'),
-          backTitle: Text('MENU'),
-        ),
-      ),
     );
   }
 }
 
-Route<dynamic> _getRoute(RouteSettings settings) {
-  if (settings.name != '/') {
-    return null;
-  }
+class _Home extends StatelessWidget {
+  const _Home();
 
-  return MaterialPageRoute<void>(
-    settings: settings,
-    builder: (context) => CraneApp(),
-    fullscreenDialog: true,
-  );
+  @override
+  Widget build(BuildContext context) {
+    return const ApplyTextOptions(
+      child: Backdrop(
+        frontLayer: SizedBox(),
+        backLayerItems: [
+          FlyForm(),
+          SleepForm(),
+          EatForm(),
+        ],
+        frontTitle: Text('CRANE'),
+        backTitle: Text('MENU'),
+      ),
+    );
+  }
 }
