@@ -9,7 +9,7 @@ import 'package:gallery/studies/crane/backlayer.dart';
 import 'package:gallery/studies/crane/header_form.dart';
 
 class FlyForm extends BackLayerItem {
-  FlyForm({int index}) : super(index: index);
+  const FlyForm({int index}) : super(index: index);
 
   @override
   _FlyFormState createState() => _FlyFormState();
