@@ -7,8 +7,6 @@ import 'package:gallery/data/gallery_options.dart';
 import 'package:gallery/l10n/gallery_localizations.dart';
 import 'package:gallery/layout/adaptive.dart';
 import 'package:gallery/layout/text_scale.dart';
-import 'package:gallery/pages/home.dart';
-import 'package:gallery/layout/focus_traversal_policy.dart';
 import 'package:gallery/studies/rally/tabs/accounts.dart';
 import 'package:gallery/studies/rally/tabs/bills.dart';
 import 'package:gallery/studies/rally/tabs/budgets.dart';
@@ -20,6 +18,8 @@ const int turnsToRotateRight = 1;
 const int turnsToRotateLeft = 3;
 
 class HomePage extends StatefulWidget {
+  const HomePage();
+
   @override
   _HomePageState createState() => _HomePageState();
 }
@@ -132,30 +132,23 @@ class _HomePageState extends State<HomePage>
         ],
       );
     }
-    final backButtonFocusNode =
-        InheritedFocusNodes.of(context).backButtonFocusNode;
-
-    return FocusTraversalGroup(
-      policy: EdgeChildrenFocusTraversalPolicy(
-        firstFocusNodeOutsideScope: backButtonFocusNode,
-        lastFocusNodeOutsideScope: backButtonFocusNode,
-        focusScope: FocusScope.of(context),
-      ),
-      child: ApplyTextOptions(
-        child: Scaffold(
-          body: SafeArea(
-            // For desktop layout we do not want to have SafeArea at the top and
-            // bottom to display 100% height content on the accounts view.
-            top: !isDesktop,
-            bottom: !isDesktop,
-            child: Theme(
-              // This theme effectively removes the default visual touch
-              // feedback for tapping a tab, which is replaced with a custom
-              // animation.
-              data: theme.copyWith(
-                splashColor: Colors.transparent,
-                highlightColor: Colors.transparent,
-              ),
+    return ApplyTextOptions(
+      child: Scaffold(
+        body: SafeArea(
+          // For desktop layout we do not want to have SafeArea at the top and
+          // bottom to display 100% height content on the accounts view.
+          top: !isDesktop,
+          bottom: !isDesktop,
+          child: Theme(
+            // This theme effectively removes the default visual touch
+            // feedback for tapping a tab, which is replaced with a custom
+            // animation.
+            data: theme.copyWith(
+              splashColor: Colors.transparent,
+              highlightColor: Colors.transparent,
+            ),
+            child: FocusTraversalGroup(
+              policy: OrderedTraversalPolicy(),
               child: tabBarView,
             ),
           ),
@@ -230,16 +223,19 @@ class _RallyTabBar extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return TabBar(
-      // Setting isScrollable to true prevents the tabs from being
-      // wrapped in [Expanded] widgets, which allows for more
-      // flexible sizes and size animations among tabs.
-      isScrollable: true,
-      labelPadding: EdgeInsets.zero,
-      tabs: tabs,
-      controller: tabController,
-      // This hides the tab indicator.
-      indicatorColor: Colors.transparent,
+    return FocusTraversalOrder(
+      order: const NumericFocusOrder(0),
+      child: TabBar(
+        // Setting isScrollable to true prevents the tabs from being
+        // wrapped in [Expanded] widgets, which allows for more
+        // flexible sizes and size animations among tabs.
+        isScrollable: true,
+        labelPadding: EdgeInsets.zero,
+        tabs: tabs,
+        controller: tabController,
+        // This hides the tab indicator.
+        indicatorColor: Colors.transparent,
+      ),
     );
   }
 }
@@ -331,7 +327,7 @@ class _RallyTabState extends State<_RallyTab>
     final unitWidth = width / (tabCount + expandedTitleWidthMultiplier);
 
     return ConstrainedBox(
-      constraints: BoxConstraints(minHeight: 56),
+      constraints: const BoxConstraints(minHeight: 56),
       child: Row(
         children: [
           FadeTransition(
