@@ -8,13 +8,14 @@ import 'package:flutter/services.dart';
 import 'package:gallery/data/gallery_options.dart';
 import 'package:gallery/l10n/gallery_localizations.dart';
 import 'package:gallery/layout/adaptive.dart';
-import 'package:gallery/layout/focus_traversal_policy.dart';
 import 'package:gallery/layout/image_placeholder.dart';
 import 'package:gallery/layout/text_scale.dart';
-import 'package:gallery/pages/home.dart';
+import 'package:gallery/studies/rally/app.dart';
 import 'package:gallery/studies/rally/colors.dart';
 
 class LoginPage extends StatefulWidget {
+  const LoginPage();
+
   @override
   _LoginPageState createState() => _LoginPageState();
 }
@@ -25,22 +26,12 @@ class _LoginPageState extends State<LoginPage> {
 
   @override
   Widget build(BuildContext context) {
-    final backButtonFocusNode =
-        InheritedFocusNodes.of(context).backButtonFocusNode;
-
-    return FocusTraversalGroup(
-      policy: EdgeChildrenFocusTraversalPolicy(
-        firstFocusNodeOutsideScope: backButtonFocusNode,
-        lastFocusNodeOutsideScope: backButtonFocusNode,
-        focusScope: FocusScope.of(context),
-      ),
-      child: ApplyTextOptions(
-        child: Scaffold(
-          body: SafeArea(
-            child: _MainView(
-              usernameController: _usernameController,
-              passwordController: _passwordController,
-            ),
+    return ApplyTextOptions(
+      child: Scaffold(
+        body: SafeArea(
+          child: _MainView(
+            usernameController: _usernameController,
+            passwordController: _passwordController,
           ),
         ),
       ),
@@ -66,7 +57,7 @@ class _MainView extends StatelessWidget {
   final TextEditingController passwordController;
 
   void _login(BuildContext context) {
-    Navigator.pop(context);
+    Navigator.of(context).pushNamed(RallyApp.homeRoute);
   }
 
   @override
@@ -95,7 +86,7 @@ class _MainView extends StatelessWidget {
       ];
     } else {
       listViewChildren = [
-        _SmallLogo(),
+        const _SmallLogo(),
         _UsernameInput(
           usernameController: usernameController,
         ),
@@ -113,7 +104,7 @@ class _MainView extends StatelessWidget {
 
     return Column(
       children: [
-        if (isDesktop) _TopBar(),
+        if (isDesktop) const _TopBar(),
         Expanded(
           child: Align(
             alignment: isDesktop ? Alignment.center : Alignment.topCenter,
@@ -140,7 +131,7 @@ class _TopBar extends StatelessWidget {
     return Container(
       width: double.infinity,
       margin: const EdgeInsets.only(top: 8),
-      padding: EdgeInsets.symmetric(horizontal: 30),
+      padding: const EdgeInsets.symmetric(horizontal: 30),
       child: Wrap(
         alignment: WrapAlignment.spaceBetween,
         children: [
@@ -151,7 +142,8 @@ class _TopBar extends StatelessWidget {
                 child: SizedBox(
                   height: 80,
                   child: FadeInImagePlaceholder(
-                    image: AssetImage('logo.png', package: 'rally_assets'),
+                    image:
+                        const AssetImage('logo.png', package: 'rally_assets'),
                     placeholder: LayoutBuilder(builder: (context, constraints) {
                       return SizedBox(
                         width: constraints.maxHeight,
@@ -197,8 +189,8 @@ class _SmallLogo extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(vertical: 64),
+    return const Padding(
+      padding: EdgeInsets.symmetric(vertical: 64),
       child: SizedBox(
         height: 160,
         child: ExcludeSemantics(
@@ -384,7 +376,7 @@ class _BorderButton extends StatelessWidget {
       ),
       textColor: Colors.white,
       onPressed: () {
-        Navigator.pop(context);
+        Navigator.of(context).pushNamed(RallyApp.homeRoute);
       },
       child: Text(text),
     );
