@@ -19,9 +19,10 @@ import 'package:gallery/studies/shrine/theme.dart';
 import 'package:scoped_model/scoped_model.dart';
 
 class ShrineApp extends StatefulWidget {
-  const ShrineApp({Key key, this.navigatorKey}) : super(key: key);
+  const ShrineApp();
 
-  final GlobalKey<NavigatorState> navigatorKey;
+  static const String loginRoute = '/shrine/login';
+  static const String homeRoute = '/shrine';
 
   @override
   _ShrineAppState createState() => _ShrineAppState();
@@ -72,7 +73,7 @@ class _ShrineAppState extends State<ShrineApp> with TickerProviderStateMixin {
   }
 
   Widget desktopBackdrop() {
-    return DesktopBackdrop(
+    return const DesktopBackdrop(
       frontLayer: ProductPage(),
       backLayer: CategoryMenuPage(),
     );
@@ -95,32 +96,41 @@ class _ShrineAppState extends State<ShrineApp> with TickerProviderStateMixin {
     final bool isDesktop = isDisplayDesktop(context);
 
     final Widget backdrop = isDesktop ? desktopBackdrop() : mobileBackdrop();
+    final Widget home = LayoutCache(
+      layouts: _layouts,
+      child: PageStatus(
+        menuController: _controller,
+        cartController: _expandingController,
+        child: HomePage(
+          backdrop: backdrop,
+          scrim: Scrim(controller: _expandingController),
+          expandingBottomSheet: ExpandingBottomSheet(
+            hideController: _controller,
+            expandingController: _expandingController,
+          ),
+        ),
+      ),
+    );
 
     return ScopedModel<AppStateModel>(
       model: _model,
       child: WillPopScope(
         onWillPop: _onWillPop,
         child: MaterialApp(
-          navigatorKey: widget.navigatorKey,
           title: 'Shrine',
           debugShowCheckedModeBanner: false,
-          home: LayoutCache(
-            layouts: _layouts,
-            child: PageStatus(
-              menuController: _controller,
-              cartController: _expandingController,
-              child: HomePage(
-                backdrop: backdrop,
-                scrim: Scrim(controller: _expandingController),
-                expandingBottomSheet: ExpandingBottomSheet(
-                  hideController: _controller,
-                  expandingController: _expandingController,
-                ),
+          initialRoute: ShrineApp.loginRoute,
+          onGenerateInitialRoutes: (_) {
+            return [
+              MaterialPageRoute<void>(
+                builder: (context) => const LoginPage(),
               ),
-            ),
-          ),
-          initialRoute: '/login',
-          onGenerateRoute: _getRoute,
+            ];
+          },
+          routes: {
+            ShrineApp.loginRoute: (context) => const LoginPage(),
+            ShrineApp.homeRoute: (context) => home,
+          },
           theme: shrineTheme.copyWith(
             platform: GalleryOptions.of(context).platform,
           ),
@@ -132,16 +142,4 @@ class _ShrineAppState extends State<ShrineApp> with TickerProviderStateMixin {
       ),
     );
   }
-}
-
-Route<dynamic> _getRoute(RouteSettings settings) {
-  if (settings.name != '/login') {
-    return null;
-  }
-
-  return MaterialPageRoute<void>(
-    settings: settings,
-    builder: (context) => LoginPage(),
-    fullscreenDialog: true,
-  );
 }
