@@ -10,29 +10,21 @@ import 'package:gallery/pages/backdrop.dart';
 
 void main() {
   testWidgets('Home page hides settings semantics when closed', (tester) async {
-    final animationController = AnimationController(
-      duration: Duration(milliseconds: 1),
-      vsync: const TestVSync(),
-    );
-    final isSettingsOpen = ValueNotifier(false);
     await tester.pumpWidget(
       MaterialApp(
-        localizationsDelegates: [GalleryLocalizations.delegate],
+        localizationsDelegates: const [GalleryLocalizations.delegate],
         home: ModelBinding(
-          initialModel: GalleryOptions(
+          initialModel: const GalleryOptions(
             textScaleFactor: 1.0,
           ),
-          child: Backdrop(
-            frontLayer: Text('Front'),
-            backLayer: Text('Back'),
-            controller: animationController,
-            isSettingsOpenNotifier: isSettingsOpen,
-            openSettingsAnimation: animationController,
+          child: const Backdrop(
+            settingsPage: Text('Front'),
+            homePage: Text('Back'),
           ),
         ),
       ),
     );
-    await tester.pump(Duration(seconds: 1));
+    await tester.pump(const Duration(seconds: 1));
 
     expect(find.bySemanticsLabel('Settings'), findsOneWidget);
     expect(find.bySemanticsLabel('Close settings'), findsNothing);
@@ -40,7 +32,7 @@ void main() {
     expect(tester.getSemantics(find.text('Front')).label, '');
 
     await tester.tap(find.bySemanticsLabel('Settings'));
-    await tester.pump(Duration(seconds: 1));
+    await tester.pump(const Duration(seconds: 1));
 
     // The test no longer finds Setting and Close settings since the semantics
     // are excluded when settings mode is activated.
@@ -48,7 +40,5 @@ void main() {
     // bottom by utilizing an invisible widget within the Settings Page.
     expect(tester.getSemantics(find.text('Back')).owner, null);
     expect(tester.getSemantics(find.text('Front')).label, 'Front');
-
-    animationController.dispose();
   });
 }
